Fix pdf download violations
