+++ frontend/src/metabase/dashboard/components/DashboardHeader/buttons/ExportAsPdfButton.unit.spec.tsx
@@ -8,10 +8,12 @@ import { createMockDashboard } from "metabase-types/api/mocks";
 import { ExportAsPdfButton } from "./ExportAsPdfButton";
 
 jest.mock("metabase/visualizations/lib/save-dashboard-pdf", () => ({
+  ...jest.requireActual("metabase/visualizations/lib/save-dashboard-pdf"),
   saveDashboardPdf: jest.fn(() => new Promise(() => {})),
 }));
 
-jest.mock("metabase/dashboard/analytics", () => ({
+jest.mock("metabase/redux/analytics", () => ({
+  ...jest.requireActual("metabase/redux/analytics"),
   trackExportDashboardToPDF: jest.fn(),
 }));
