+++ frontend/src/metabase/dashboard/constants.ts
@@ -49,11 +49,6 @@ export const INITIAL_DASHBOARD_STATE: DashboardState = {
 
 export const DASHBOARD_SLOW_TIMEOUT = 15 * 1000;
 
-export const DASHBOARD_PDF_EXPORT_ROOT_ID =
-  "Dashboard-Parameters-And-Cards-Container";
-export const DASHBOARD_HEADER_PARAMETERS_PDF_EXPORT_NODE_ID =
-  "Dashboard-Parameters-Content";
-
 export const DEFAULT_DASHBOARD_DISPLAY_OPTIONS: EmbedDisplayParams = {
   background: true,
   bordered: false,
