+++ frontend/src/metabase/redux/analytics.ts
@@ -1,8 +1,28 @@
 import { trackSchemaEvent } from "metabase/analytics";
+import type { DashboardId } from "metabase-types/api";
 
 import type { ResourceAccessedVia, ResourceType } from "./downloads";
 
 const SCHEMA = "downloads";
+
+// dashboard_id is required in the snowplow schema, but we don't send UUIDs or
+// JWTs in public/static-embed scenarios, so fall back to 0 there.
+const getTrackedDashboardId = (dashboardId: DashboardId | undefined) =>
+  typeof dashboardId === "number" ? dashboardId : 0;
+
+export const trackExportDashboardToPDF = ({
+  dashboardId,
+  dashboardAccessedVia,
+}: {
+  dashboardId?: DashboardId;
+  dashboardAccessedVia: ResourceAccessedVia;
+}) => {
+  trackSchemaEvent("dashboard", {
+    event: "dashboard_pdf_exported",
+    dashboard_id: getTrackedDashboardId(dashboardId),
+    dashboard_accessed_via: dashboardAccessedVia,
+  });
+};
 
 export const trackDownloadResults = ({
   resourceType,
