+++ frontend/src/metabase/visualizations/lib/save-dashboard-pdf.ts
@@ -14,6 +14,12 @@ import {
 import { fixParameterLegendOffsetForExport } from "./image-exports";
 import { SAVING_DOM_IMAGE_CLASS } from "./save-chart-image";
 
+// DOM ids on exportable nodes so the PDF exporter and downloads thunk can find them
+export const DASHBOARD_PDF_EXPORT_ROOT_ID =
+  "Dashboard-Parameters-And-Cards-Container";
+export const DASHBOARD_HEADER_PARAMETERS_PDF_EXPORT_NODE_ID =
+  "Dashboard-Parameters-Content";
+
 const TARGET_ASPECT_RATIO = 21 / 17;
 
 interface DashCardBounds {
